Bugfixes for the Synchronization Prozess

+++ OwncloudUniversal.Shared/WebDav/DavItem.cs
@@ -12,6 +12,6 @@ namespace OwncloudUniversal.Shared.WebDav
         public string ContentType { get; set; }
         public DateTime? LastModified { get; set; }
         public string DisplayName { get; set; }
-        public int? ContentLength { get; set; }
+        public long? ContentLength { get; set; }
     }
 }
